careerai/backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
motor==3.4.0
pymongo==4.7.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.2.1

# PDF Extraction
pdfplumber==0.11.0
PyPDF2==3.0.1

# DOCX
python-docx==1.1.2

# NLP
spacy==3.7.4
nltk==3.8.1

# ML
scikit-learn==1.4.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
